removed duplicate build_test_env fixture.
